+++ common/net_table.go
@@ -12,17 +12,17 @@ import (
 type NetTable struct {
 	BaseService
 
-	localNode *LocalNode
-	waitGroup sync.WaitGroup
-	dhtInChan chan string
+	localNode       *LocalNode
+	waitGroup       sync.WaitGroup
+	dhtInChan       chan string
 
-	lock      sync.RWMutex
-	blackList map[string]time.Time
-	peers     map[string]*RemoteNode
+	lock            sync.RWMutex
+	blackList       map[string]time.Time
+	peers           map[string]*RemoteNode
 
 	heartbeatTicker <-chan time.Time
 
-	logger log.Logger
+	logger          log.Logger
 }
 
 func (nt NetTable) Name() string {
@@ -55,7 +55,7 @@ func (nt *NetTable) Stop() {
 	}
 }
 
-func (nt *NetTable) GetDHTInChannel() chan<- string {
+func (nt *NetTable) GetDHTInChannel() chan <- string {
 	return nt.dhtInChan
 }
 
@@ -107,7 +107,9 @@ func (nt *NetTable) processDHTIn() {
 			nt.lock.Unlock()
 
 			if !ok {
-				nt.tryConnect(host)
+				if err := nt.tryConnect(host); err != nil {
+					nt.logger.Info("unable connect %s err: %s", host, err)
+				}
 			}
 		}
 	}
@@ -133,14 +135,15 @@ func (nt *NetTable) heartbeat() {
 	}
 }
 
-func (nt *NetTable) tryConnect(h string) {
+func (nt *NetTable) tryConnect(h string) error {
 	rn, err := TryConnect(h, nt.localNode.NetworkSecret(), nt.localNode)
 	if err != nil {
 		nt.addToBlackList(h)
-		return
+		return err
 	}
 	nt.logger.Debug("adding remote node...")
 	nt.AddRemoteNode(rn)
+	return nil
 }
 
 func (nt *NetTable) addToBlackList(h string) {
@@ -154,9 +157,8 @@ func (nt *NetTable) SendPacket(dstIP net.IP, payload []byte) {
 
 	rn := nt.RemoteNodeByIP(dstIP)
 	if rn == nil {
-		nt.logger.Error("destination host %q unreachable", dstIP.String())
+		nt.logger.Debug("destination host %q unreachable", dstIP.String())
 		nt.logger.Debug("known hosts, %v", nt.knownHosts())
-
 		return
 	}
